mcp_server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-coscientist-mcp-server"
version = "0.1.0"
description = "MCP server for Open Coscientist literature review"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Jataware Corp", email = "info@jataware.com"}
]
license = {text = "Proprietary"}
keywords = ["mcp", "pubmed", "literature-review"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp~=2.14.3",
    "fastapi~=0.128.0",
    "httpx~=0.28.1",
    "beautifulsoup4~=4.14.3",
    "lxml~=6.0.2",
    "biopython~=1.86",
    "python-dotenv~=1.2.1",
    "uvicorn~=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Repository = "https://github.com/jataware/open-coscientist"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_server*"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
